+++ railties/lib/rails/application.rb
@@ -485,13 +485,14 @@ module Rails
     # +production+ environment), or +config/credentials.yml.enc+ if that file
     # does not exist.
     #
-    # The encryption key is taken from either <tt>ENV["RAILS_MASTER_KEY"]</tt>,
+    # The encryption key is taken from either the env variable +config.credentials.env_key+,
     # or from the file specified by +config.credentials.key_path+. By default,
+    # +config.credentials.env_key+ will pull from <tt>ENV["RAILS_MASTER_KEY"]</tt> and
     # +config.credentials.key_path+ will point to either
     # <tt>config/credentials/#{environment}.key</tt> for the current
     # environment, or +config/master.key+ if that file does not exist.
     def credentials
-      @credentials ||= encrypted(config.credentials.content_path, key_path: config.credentials.key_path)
+      @credentials ||= encrypted(config.credentials.content_path, key_path: config.credentials.key_path, env_key: config.credentials.env_key)
     end
 
     # Returns an ActiveSupport::EncryptedConfiguration instance for an encrypted